修复23041414引入的错误

+++ utilities/master-ctrl/src/ax_uart.c
@@ -2375,12 +2375,12 @@ void parse_send_fw_info_msg(void)
 
             if ((updateFw.fileSize % 128) == 0)
             {
-                updateFw.fwMsgSumConform = (uint8_t)(updateFw.fileSize / 128);
+                updateFw.fwMsgSumConform = (uint16_t)(updateFw.fileSize / 128);
             }
 
             else
             {
-                updateFw.fwMsgSumConform = (uint8_t)(updateFw.fileSize / 128 + 1);
+                updateFw.fwMsgSumConform = (uint16_t)(updateFw.fileSize / 128 + 1);
             }
 
             if ((updateFw.fileSize > APP_FW_MAX_LEN) || (updateFw.fwMsgSum != updateFw.fwMsgSumConform))
